sub_section for json
Rethink of split_path: Split is the method to create a path+name that can be found in table.
So the previous logic, that treated ending / and a delimiter does not make sense. I had a feeling that it was wrong, now I know. So now: /a/b/ results in '/a/' + 'b', and / results in '' + '' which is existing root element !
This however allows acessing attrs with '/' at the end of the path - like '/value/'. Maybe this is not good. But if needed this can be guarded in `get_item` instead of `split_path`