Api shape unification (#19)
* sub_section for json

Rethink of split_path: Split is the method to create a path+name that can be found in table.
So the previous logic, that treated ending / and a delimiter does not make sense. I had a feeling that it was wrong, now I know. So now: /a/b/ results in '/a/' + 'b', and / results in '' + '' which is existing root element !
This however allows acessing attrs with '/' at the end of the path - like '/value/'. Maybe this is not good. But if needed this can be guarded in `get_item` instead of `split_path`

* change ping to GET

* support new api shape for check_object

* linter fixes

* more linter fixes

* linter ... mmmm

* get the linter fix back to original

* Update zcl_abaplint_json_reader.clas.abap

* Update zcl_abaplint_json_reader.clas.abap

Co-authored-by: Lars Hvam <larshp@hotmail.com>