New feature: Diff for abaplint configuration
It's hard to tell if your local config is different from the config found in the corresponding repo. Similarly, it's hard to determine if there are new or changed abaplint checks available.

The PR adds two buttons to the configuration screen to display the diff between the current config and either default or repo configuration.